backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
PyJWT
tensorflow
pandas
openpyxl
scikit-learn